reduce duplicated code
