+++ packages/suite-base/src/panels/RawMessagesCommon/DiffStats.tsx
